WebApp/requirements.txt
flask
requests
pillow
pymongo